refactor: RandomNumberGenerator와 구현체를 domain 패키지로 이동
